+++ fabric/src/main/java/org/thinkingstudio/ryoamiclights/fabric/RyoamicLightsCompatImpl.java
@@ -2,6 +2,8 @@ package org.thinkingstudio.ryoamiclights.fabric;
 
 import net.fabricmc.loader.api.FabricLoader;
 
+import java.nio.file.Path;
+
 public class RyoamicLightsCompatImpl {
     public static boolean isCanvasInstalled() {
         return FabricLoader.getInstance().isModLoaded("canvas");
@@ -21,5 +23,9 @@ public class RyoamicLightsCompatImpl {
 
     public static boolean isDevEnvironment() {
         return FabricLoader.getInstance().isDevelopmentEnvironment();
+    }
+
+    public static Path getConfigDir() {
+        return FabricLoader.getInstance().getConfigDir();
     }
 }