+++ neoforge/src/main/java/org/thinkingstudio/ryoamiclights/neoforge/RyoamicLightsNeoForge.java
@@ -10,38 +10,48 @@
 
 package org.thinkingstudio.ryoamiclights.neoforge;
 
-import dev.architectury.platform.Platform;
 import net.minecraft.client.MinecraftClient;
+import net.minecraft.resource.SynchronousResourceReloader;
 import net.neoforged.bus.api.EventPriority;
+import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.IExtensionPoint;
 import net.neoforged.fml.ModLoadingContext;
 import net.neoforged.fml.loading.FMLLoader;
+import net.neoforged.neoforge.client.ConfigScreenHandler;
+import net.neoforged.neoforge.client.event.RegisterClientReloadListenersEvent;
 import net.neoforged.neoforge.client.event.RenderLevelStageEvent;
 import net.neoforged.neoforge.common.NeoForge;
 import org.thinkingstudio.ryoamiclights.RyoamicLights;
 import net.neoforged.fml.common.Mod;
+import org.thinkingstudio.ryoamiclights.api.item.ItemLightSources;
 import org.thinkingstudio.ryoamiclights.gui.SettingsScreen;
 import org.thinkingstudio.ryoamiclights.neoforge.api.DynamicLightsInitializerEvent;
 
 @Mod(RyoamicLights.NAMESPACE)
 public class RyoamicLightsNeoForge {
-    public RyoamicLightsNeoForge() {
+    public RyoamicLightsNeoForge(IEventBus modEventBus) {
         if (FMLLoader.getDist().isClient()) {
-            this.onInitializeClient();
+            this.onInitializeClient(modEventBus);
         }
     }
 
-    public void onInitializeClient() {
-        new RyoamicLights().clientInit();
-        ModLoadingContext.get().registerExtensionPoint(IExtensionPoint.DisplayTest.class, () -> new IExtensionPoint.DisplayTest(() -> IExtensionPoint.DisplayTest.IGNORESERVERONLY, (a, b) -> true));
-        NeoForge.EVENT_BUS.addListener(EventPriority.HIGHEST, RenderLevelStageEvent.class, event -> {
+    public void onInitializeClient(IEventBus modEventBus) {
+        ModLoadingContext context = ModLoadingContext.get();
+
+        RyoamicLights.get().clientInit();
+
+        context.registerExtensionPoint(IExtensionPoint.DisplayTest.class, () -> new IExtensionPoint.DisplayTest(() -> IExtensionPoint.DisplayTest.IGNORESERVERONLY, (a, b) -> true));
+        context.registerExtensionPoint(ConfigScreenHandler.ConfigScreenFactory.class, () -> new ConfigScreenHandler.ConfigScreenFactory((client, screen) -> new SettingsScreen(screen)));
+        modEventBus.addListener(EventPriority.HIGHEST, RenderLevelStageEvent.class, event -> {
             if (event.getStage() == RenderLevelStageEvent.Stage.AFTER_TRIPWIRE_BLOCKS) {
                 MinecraftClient.getInstance().getProfiler().swap("dynamic_lighting");
                 RyoamicLights.get().updateAll(event.getLevelRenderer());
             }
         });
-        NeoForge.EVENT_BUS.post(new DynamicLightsInitializerEvent());
+        modEventBus.addListener(EventPriority.HIGHEST, RegisterClientReloadListenersEvent.class, event -> {
+            event.registerReloadListener((SynchronousResourceReloader) ItemLightSources::load);
+        });
 
-        Platform.getMod(RyoamicLights.NAMESPACE).registerConfigurationScreen(SettingsScreen::new);
+        NeoForge.EVENT_BUS.post(new DynamicLightsInitializerEvent());
     }
 }