+++ fabric/CHANGELOG.md
@@ -1,2 +1,3 @@
 ## Changelog
-- code clean
+- remove the mixin for embeddium (only forge)
+- remove architectury dependency